requirements.txt
PySide6
requests
python-dotenv
pyyaml
google-generativeai
websocket-client
pygame
google-cloud-texttospeech